+++ src/components/atoms/WorkList.tsx
@@ -1,5 +1,5 @@
-import { Typography } from "@material-ui/core";
 import React from "react";
+import WorkItem from "src/components/atoms/WorkItem";
 import { Work } from "src/graphQL/type";
 import styled from "styled-components";
 
@@ -39,49 +39,6 @@ export default (
     </WorkList>
 );
 
-interface WorkItemProps extends React.HTMLAttributes<HTMLDivElement> {
-    work: Work;
-}
-
-const WorkItem = (
-    {
-        work,
-        ...props
-    }: WorkItemProps
-) => (
-    <WorkItemBase
-        {...props}
-        unselectable={undefined}
-    >
-        <WorkImage
-            src={(
-                work.imageUrl ? work.imageUrl
-              :                 "/img/no-image.png"
-            )}
-        />
-        <div>
-            <Typography variant="caption">
-                {work.user && work.user.displayName}
-            </Typography>
-            <Typography gutterBottom variant="h6" component="h2">
-                {work.title}
-            </Typography>
-        </div>
-    </WorkItemBase>
-);
-
-const WorkImage = styled.img`
-    width: 100%;
-    border-radius: 8px;
-    transition: all 0.15s ease-in-out;
-    cursor: pointer;
-    :hover {
-        background-color: #fff;
-        transform: scale(1.1);
-        box-shadow: 0 7px 14px rgba(0,0,0,0.25), 0 5px 5px rgba(0,0,0,0.22);
-    }
-`;
-
 const WorkList = styled.div`
     margin: 0 3rem;
     display: flex;
@@ -91,9 +48,4 @@ const WorkList = styled.div`
         align-items: center;
         margin: 0 1rem;
     }
-`;
-
-const WorkItemBase = styled.div`
-    margin-bottom: 2rem;
-    width: 100%;
 `;
